pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Experiments for Master Thesis"
authors = [
    {name = "Lisa Schmierer",email = "lisa.schmierer@tum.de"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "numpy (<2.3.0,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scapy (>=2.6.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pickl (>=0.3.0,<0.4.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "dpkt (>=1.9.8,<2.0.0)",
    "snakeviz (>=2.2.2,<3.0.0)",
    "xxhash (>=3.5.0,<4.0.0)",
    "numba (>=0.61.2,<0.62.0)",
    "bottleneck (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
